chore: remove derived-properties.d.ts
